[bug] commit with bugs in create right of way
